services/python-collector/requirements.txt
APScheduler==3.11.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
geographiclib==2.1
geopy==2.4.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pika==1.3.2
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
PyYAML==6.0.3
requests==2.32.5
schedule==1.2.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
python-dotenv==1.0.1